+++ R/tidy.R
@@ -780,7 +780,7 @@ get_re_tidy_list <- function(x, crit, model = 1, delta = FALSE) {
     re_b_df <- re_b_df[, c("group_name", "term", "level_ids", "estimate", "std.error", "conf.low", "conf.high")]
   }
   # remove ":" in the level_ids
-  re_b_df$level_ids <- sapply(strsplit(re_b_df$level_ids, ":"), function(x) x[2])
+  re_b_df$level_ids <- sapply(strsplit(re_b_df$level_ids, ":"), function(x) x[3])
   out_ranef <- re_b_df
   row.names(out_ranef) <- NULL
 
@@ -800,6 +800,19 @@ get_re_tidy_list <- function(x, crit, model = 1, delta = FALSE) {
     indx <- which(group_key$model == re_cov_df$model[i] & group_key$group_id == (re_cov_df$group_indices[i] + 1))
     re_cov_df$group[i] <- group_key$group_name[indx]
   }
+
+  # Uncorrelated random intercepts and slopes for the same group, e.g. (1|group)+(0+slope|group),
+  # are fit internally as if they were separate groups, but their SDs should be displayed
+  # in a 2-by-2 matrix with NAs on the off-diagonals.
+  true_group_dfs <- list()
+  for (i in seq_len(length(unique(re_cov_df$group)))) {
+    group <- unique(re_cov_df$group)[i]
+    true_group_dfs[[i]] <- re_cov_df[re_cov_df$group == group,]
+    correction <- true_group_dfs[[i]]$group_indices - true_group_dfs[[i]]$group_indices[1]
+    true_group_dfs[[i]]$rows <- true_group_dfs[[i]]$rows + correction
+    true_group_dfs[[i]]$cols <- true_group_dfs[[i]]$cols + correction
+  }
+  re_cov_df <- do.call(rbind, true_group_dfs)
 
   re_indx <- grep("re_cov_pars", names(x$sd_report$value), fixed = TRUE)
   non_nas <- which(x$sd_report$value[re_indx] != 0) # remove parameter that gets mapped off